backend/requirements.txt
Django>=5.0
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0
django-cors-headers>=4.3.1
django-filter>=23.5
djangorestframework>=3.14.0
djangorestframework-gis==1.0
djangorestframework-simplejwt>=5.3.1
drf-yasg==1.21.7
Pillow>=10.1.0
psycopg2>=2.9.9
python-dotenv>=1.0.0
redis>=5.0.1
django-redis>=5.4.0
scipy==1.12.0
numpy==1.26.4
requests>=2.31.0
# Additional common dependencies
pandas>=2.2.0
pytz>=2024.1
celery>=5.3.6
gunicorn>=21.2.0
whitenoise>=6.6.0
django-storages>=1.14.2
boto3>=1.34.7
django-environ>=0.11.2
Faker>=22.5.0 